+++ src/client/components/dt_count_badge/dt_count_badge.js
@@ -5,6 +5,7 @@
  * 
  * Data context:
  * - item: the item as provided to the tabular display (i.e. a modified closest record)
+ *   with DYN { analyze, entity, records }
  * - table: the Tabular.Table instance
  */
 
@@ -15,14 +16,14 @@ import './dt_count_badge.html';
 Template.dt_count_badge.helpers({
     // the records count
     count(){
-        return this.item.DYN.count;
+        return this.item.DYN.records.length;
     },
     // make the badge transparent if count is just one
     classes(){
-        return this.item.DYN.count === 1 ? 'ui-transparent' : '';
+        return this.item.DYN.records.length === 1 ? 'ui-transparent' : '';
     },
     // a title to be provided if visible
     title(){
-        return this.item.DYN.count === 1 ? '' : pwixI18n.label( I18N, 'buttons.count_badge_title', this.item.label, this.item.DYN.count );
+        return this.item.DYN.records.length === 1 ? '' : pwixI18n.label( I18N, 'buttons.count_badge_title', this.item.label, this.item.DYN.count );
     }
 });